requirements.txt
# WhatsApp AI Second Brain Assistant Dependencies
# Latest compatible versions as of June 2025

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
requests==2.32.3
langchain==0.2.5
langchain-community==0.2.5
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pymupdf==1.24.5
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
twilio==9.2.0
apscheduler==3.10.4
dateparser==1.2.0
pytesseract==0.3.10
pymongo==4.7.3
motor==3.4.0
numpy==1.26.4
Pillow==10.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
